Use Github actions for releases.
This setup will build wheels for all supported platforms.
It will publish a new version on PyPi and push the package to crates.io
when we create a new tag.
With this, we also move from Pipenv to poetry, which
provides a similar config file like cargo to manage dependencies.